+++ src/constants/index.ts
@@ -3,3 +3,4 @@ export * from './DASHBOARD_LIST';
 export * from './JIRA_NOVA';
 export * from './CONFERENCE_REEL';
 export * from './TREVOR_TEAM';
+export * from './JIRA_TREVOR';