+++ src/stores/index.ts
@@ -1,2 +1,3 @@
 export * from './dashboardStore';
 export * from './jiraNovaStore';
+export * from './trevorJiraStore';